back to u32, safer

+++ crates/mimic/src/data/executor/load/dynamic.rs
@@ -89,8 +89,8 @@ impl LoadExecutorDyn {
             .load(&resolved_entity, &query.selector, None)?
             .into_iter()
             .filter(|row| include_row(row, &query, E::PATH))
-            .skip(query.offset)
-            .take(query.limit.unwrap_or(usize::MAX))
+            .skip(query.offset as usize)
+            .take(query.limit.unwrap_or(u32::MAX) as usize)
             .collect();
 
         Ok(LoadCollectionDyn(rows))
